DevMaster Debugger/devnarrative/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devnarrative"
version = "0.1.0"
description = "Transform git history into compelling human stories"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "DevNarrative Team"}
]
dependencies = [
    "click>=8.1.7",
    "rich>=13.7.0",
    "GitPython>=3.1.40",
    "pydantic>=2.5.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.12.0",
    "ruff>=0.1.8",
]

[project.scripts]
devnarrative = "devnarrative.cli:main"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
